Time: 5 ms (29.8%), Space: 45.8 MB (68.07%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -14,9 +14,9 @@ class Solution {
        int len=arr.length;
        int mid=len/2;
        if(len%2!=0){
-           return arr[mid];
+           return (double)arr[mid];
        }else {
-           return (arr[mid]+arr[mid-1])/2;
+           return (double)(arr[mid]+arr[mid-1])/2;
        }
     }
 }
